debug

+++ python/sglang/srt/models/qwen3_5.py
@@ -1012,7 +1012,7 @@ class Qwen3_5MoeForCausalLM(Qwen3_5ForCausalLM):
                         weight_loader = getattr(
                             param, "weight_loader", default_weight_loader
                         )
-                        weight_loader(param, loaded_weight)
+                        weight_loader(param, loaded_weight.to(dtype=param.dtype))
                     else:
                         logger.warning(f"Parameter {name} not found in params_dict")
             loaded_params.add(name)
@@ -1325,7 +1325,7 @@ class Qwen3_5MoeForConditionalGeneration(Qwen3VLForConditionalGeneration):
                         weight_loader = getattr(
                             param, "weight_loader", default_weight_loader
                         )
-                        weight_loader(param, loaded_weight)
+                        weight_loader(param, loaded_weight.to(dtype=param.dtype))
                     else:
                         logger.warning(f"Parameter {name} not found in params_dict")
             loaded_params.add(name)
