+++ python/sglang/srt/models/qwen3_omni_moe.py
@@ -654,7 +654,7 @@ class Qwen3OmniMoeForConditionalGeneration(PreTrainedModel):
                         weight_loader = getattr(
                             param, "weight_loader", default_weight_loader
                         )
-                        weight_loader(param, loaded_weight)
+                        weight_loader(param, loaded_weight.to(dtype=param.dtype))
                     else:
                         logger.warning(
                             f"Loaded weight with {name=} not found in params_dict"
